refactor: alter url use-cases repositories to url folder
