+++ src/core/domain/repositories/url/create-url-repository.interface.ts
@@ -1,4 +1,4 @@
-import { UrlEntity } from '../entities/url.entity';
+import { UrlEntity } from '../../entities/url.entity';
 
 export interface CreateUrlRepositoryInterface {
   save(data: UrlEntity): Promise<UrlEntity>;
