+++ src/core/domain/repositories/url/get-url-by-shortcode-repository.interface.ts
@@ -1,4 +1,4 @@
-import { UrlEntity } from '../entities/url.entity';
+import { UrlEntity } from '../../entities/url.entity';
 
 export interface GetUrlByShortCodeRepositoryInterface {
   findByShortCode(shortCode: string): Promise<UrlEntity>;
