+++ src/core/domain/repositories/url/update-url-clickcount-repository.interface.ts
@@ -1,4 +1,4 @@
-import { UrlEntity } from '../entities/url.entity';
+import { UrlEntity } from '../../entities/url.entity';
 
 export interface UpdateUrlClickCountRepositoryInterface {
   update(data: UrlEntity): Promise<UrlEntity>;
