+++ src/core/use-cases/url/get-url-by-shortcode.service.ts
@@ -1,9 +1,9 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { GetUrlByShortCodeRepositoryInterface } from 'src/core/domain/repositories/get-url-by-shortcode-repository.interface';
+import { GetUrlByShortCodeRepositoryInterface } from 'src/core/domain/repositories/url/get-url-by-shortcode-repository.interface';
 import { GetUrlByShortCodeInterface } from 'src/core/domain/use-cases/url/get-url-by-shortcode.interface';
 import { UrlRetrievalFailedError } from '../errors/url-error';
 import { UrlEntity } from 'src/core/domain/entities/url.entity';
-import { UpdateUrlClickCountRepositoryInterface } from 'src/core/domain/repositories/update-url-clickcount-repository.interface';
+import { UpdateUrlClickCountRepositoryInterface } from 'src/core/domain/repositories/url/update-url-clickcount-repository.interface';
 
 @Injectable()
 export class GetUrlByShortCodeService implements GetUrlByShortCodeInterface {
