+++ test/src/core/use-cases/url/get-url-by-shortcode.service.spec.ts
@@ -1,11 +1,11 @@
-import { GetUrlByShortCodeRepositoryInterface } from 'src/core/domain/repositories/get-url-by-shortcode-repository.interface';
+import { GetUrlByShortCodeRepositoryInterface } from 'src/core/domain/repositories/url/get-url-by-shortcode-repository.interface';
 import { GetUrlByShortCodeInterface } from 'src/core/domain/use-cases/url/get-url-by-shortcode.interface';
 import { GetUrlByShortCodeService } from 'src/core/use-cases/url/get-url-by-shortcode.service';
 import { GetUrlByShortCodeRepositoryService } from 'src/infrastructure/database/repositories/url/get-url-by-shortcode-repository.service';
 import { DatabaseErrorHandler } from 'src/infrastructure/database/utils/db-error-handler';
 import { configureDbDriverMock } from '../../../../_mocks_/configure-db-driver-mock';
 import { UrlEntity } from 'src/core/domain/entities/url.entity';
-import { UpdateUrlClickCountRepositoryInterface } from 'src/core/domain/repositories/update-url-clickcount-repository.interface';
+import { UpdateUrlClickCountRepositoryInterface } from 'src/core/domain/repositories/url/update-url-clickcount-repository.interface';
 import { UpdateUrlClickCountRepositoryService } from 'src/infrastructure/database/repositories/url/update-url-clickcount-repository.service';
 import { UrlRetrievalFailedError } from 'src/core/use-cases/errors/url-error';
